requirements.txt
requests
imageio[ffmpeg]
imageio-ffmpeg
mss
pillow